requirements.txt
# Django core
Django>=5.0,<5.1
gunicorn>=21.0

# History tracking for audit
django-simple-history>=3.4

# Environment variables
python-dotenv>=1.0
